Fixed warnings from the Intel Compiler.  On WSJ it looks like Microsoft produce a faster exe!
git-svn-id: https://groke.otago.ac.nz/ant/ant@753 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/ant_params.h
@@ -27,7 +27,7 @@ struct ANT_ANT_params
 	double stemmer_similarity_threshold;	// threshold for rejecting stems with the thesaurus (default = 0.0)
 
 	long long sort_top_k;				// accurate rank point in the accumulator sort (in the call to sort_results_list())
-	long long trim_postings_k;			// trim the postigs lists at no fewer than k
+	long trim_postings_k;				// trim the postigs lists at no fewer than k
 	long metric;						// which metric to use (MAP, MAgP, etc)
 	long metric_n;						// so we can so P@10, p@30 and so on using one metric
 	char *assessments_filename;			// name of the file containing the assessments for the given queries
