+++ source/search_engine.h
@@ -68,7 +68,7 @@ protected:
 	ANT_compressable_integer *document_lengths;
 	double mean_document_length;
 	long long stem_buffer_length_in_bytes;
-	long long trim_postings_k;
+	long trim_postings_k;
 	long long collection_length_in_terms;
 
 public:
@@ -105,7 +105,7 @@ public:
 	void stats_text_render(void);
 	void stats_all_text_render(void);
 	void stats_add(void);
-	void set_trim_postings_k(long long what) { trim_postings_k = what; }
+	void set_trim_postings_k(long what) { trim_postings_k = what; }
 	long long get_collection_length(void) { return collection_length_in_terms; }
 
 	ANT_compressable_integer *get_decompressed_postings(char *term, ANT_search_engine_btree_leaf *term_details);
